% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bahoura.R
\name{bahoura}
\alias{bahoura}
\title{Bahoura et al. QRS detection}
\usage{
bahoura(ecg, fs)
}
\arguments{
\item{ecg}{The input single-channel input vector (raw ECG)}

\item{fs}{The frequency in Hz with which the ecg was sampled}
}
\value{
Numeric array containing the indices (sample numbers) at which the fiducial R-peaks were found
}
\description{
Detect R peaks of the QRS complex in a raw ECG record, based on the
algorithm proposed by Bahoura, Hassani & Hubin (1997)
}
\details{
This function attempts to detect ECG fiducial points in a single-channel electrocardiogram signal using the
algorithm proposed by Bahoura, Hassani & Hubin (1997), which is a simplified version of the
method proposed by Li, Zheng & Tai (1995).
}
\examples{
data(rec100)
fs <- 360
pks <- bahoura(rec100$MLII, fs)

\dontrun{
# plot first 5 seconds of data
N <- 5 * fs
plot (rec100$time[1:N], rec100$MLII[1:N], type = "l", main = "MIT-BIH database, record 100",
      xlab = "Time (s)", ylab = "Amplitude (mV)")
points (pks[which(pks<=N)]/fs, rec100$MLII[pks[which(pks<=N)]], col="red")
}
}
\references{
Bahoura, M., Hassani, M., & Hubin, M. (1997). DSP implementation of wavelet transform for real time ECG
            wave forms detection and heart rate analysis. Computer Methods and Programs in Biomedicine, 52, 35-44.
            DOI: \href{https://doi.org/10.1016/S0169-2607(97)01780-X}{10.1016/S0169-2607(97)01780-X}.\cr
            Li, C., Zheng, C., & Tai, C. (1995). Detection of ECG Characteristic Points Using Wavelet Transforms.
            IEEE Transactions on Biomedical Engineering, 42(1), 21-28.
            DOI: \href{https://doi.org/10.1109/10.362922}{10.1109/10.362922}.
}
\author{
Geert van Boxtel
}
