% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRSdetect-internal.R
\name{is.scalar}
\alias{is.scalar}
\alias{is.posscal}
\alias{is.wholenumber}
\alias{RMS}
\title{QRSdectect-internal}
\usage{
is.scalar(x)

is.posscal(x)

is.wholenumber(x, tol = .Machine$double.eps^0.5)

RMS(x)
}
\description{
Internal or barely commented functions not exported from the namespace.
}
\details{
\preformatted{
## TYPE CHECKING AND CONVERSION
  is.scalar(x)                                      # test if x is a scalar
  is.posscal(x)                                     # test if x is a positive scalar
  is.wholenumber(x, tol = .Machine$double.eps^0.5)  # test if x is a whole number
  }
}
\author{
Geert van Boxtel
}
\keyword{internal}
