% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRSdetect-internal.R
\name{QRSdetect-internal}
\alias{QRSdetect-internal}
\alias{is.scalar}
\alias{is.posscal}
\alias{is.wholenumber}
\alias{RMS}
\title{QRSdetect-internal}
\usage{
is.scalar(x)

is.posscal(x)

is.wholenumber(x, tol = .Machine$double.eps^0.5)

RMS(x)
}
\arguments{
\item{x}{object}

\item{tol}{tolerance}
}
\description{
Internal or barely commented functions not exported from the namespace.
}
\author{
Geert van Boxtel
}
